Update step size on Transcendent Power input
In certain browsers (I tested in Firefox), the TP input box will gain a red border if a decimal is entered.  This is because the default step size for a number input is 1, so Firefox thinks any decimal is an invalid input.  This can be fixed by specifying the step size to .01, which also means that the arrows in the box can be used for fine-grain adjustment.

As far as I know this is an entirely cosmetic change - the calculator will work either way.